+++ wordless/wl_concordancer.py
@@ -226,21 +226,21 @@ class Wrapper_Concordancer(wl_layouts.Wl_Wrapper):
         self.group_box_table_settings = QGroupBox(self.tr('Table Settings'), self)
 
         (
-            self.checkbox_show_pct,
-            self.checkbox_show_cumulative,
-            self.checkbox_show_breakdown
+            self.checkbox_show_pct_data,
+            self.checkbox_show_cum_data,
+            self.checkbox_show_breakdown_file
         ) = wl_widgets.wl_widgets_table_settings(
             self,
             tables = [self.table_concordancer]
         )
 
-        self.checkbox_show_cumulative.hide()
-        self.checkbox_show_breakdown.hide()
+        self.checkbox_show_cum_data.hide()
+        self.checkbox_show_breakdown_file.hide()
 
-        self.checkbox_show_pct.stateChanged.connect(self.table_settings_changed)
+        self.checkbox_show_pct_data.stateChanged.connect(self.table_settings_changed)
 
         self.group_box_table_settings.setLayout(wl_layouts.Wl_Layout())
-        self.group_box_table_settings.layout().addWidget(self.checkbox_show_pct, 0, 0)
+        self.group_box_table_settings.layout().addWidget(self.checkbox_show_pct_data, 0, 0)
 
         # Figure Settings
         self.group_box_fig_settings = QGroupBox(self.tr('Figure Settings'), self)
@@ -341,7 +341,7 @@ class Wrapper_Concordancer(wl_layouts.Wl_Wrapper):
         self.combo_box_width_unit.setCurrentText(settings['generation_settings']['width_unit'])
 
         # Table Settings
-        self.checkbox_show_pct.setChecked(settings['table_settings']['show_pct'])
+        self.checkbox_show_pct_data.setChecked(settings['table_settings']['show_pct_data'])
 
         # Figure Settings
         self.combo_box_sort_results_by.setCurrentText(settings['fig_settings']['sort_results_by'])
@@ -429,7 +429,7 @@ class Wrapper_Concordancer(wl_layouts.Wl_Wrapper):
     def table_settings_changed(self):
         settings = self.main.settings_custom['concordancer']['table_settings']
 
-        settings['show_pct'] = self.checkbox_show_pct.isChecked()
+        settings['show_pct_data'] = self.checkbox_show_pct_data.isChecked()
 
     def fig_settings_changed(self):
         settings = self.main.settings_custom['concordancer']['fig_settings']
@@ -582,7 +582,7 @@ class Wl_Table_Concordancer(wl_tables.Wl_Table_Data_Sort_Search):
 
                 self.enable_updates()
 
-                self.toggle_pct()
+                self.toggle_pct_data()
             except Exception:
                 err_msg = traceback.format_exc()
             finally: