+++ wordless/wl_dependency_parser.py
@@ -150,21 +150,21 @@ class Wrapper_Dependency_Parser(wl_layouts.Wl_Wrapper):
         self.group_box_table_settings = QGroupBox(self.tr('Table Settings'), self)
 
         (
-            self.checkbox_show_pct,
-            self.checkbox_show_cumulative,
-            self.checkbox_show_breakdown
+            self.checkbox_show_pct_data,
+            self.checkbox_show_cum_data,
+            self.checkbox_show_breakdown_file
         ) = wl_widgets.wl_widgets_table_settings(
             self,
             tables = [self.table_dependency_parser]
         )
 
-        self.checkbox_show_cumulative.hide()
-        self.checkbox_show_breakdown.hide()
+        self.checkbox_show_cum_data.hide()
+        self.checkbox_show_breakdown_file.hide()
 
-        self.checkbox_show_pct.stateChanged.connect(self.table_settings_changed)
+        self.checkbox_show_pct_data.stateChanged.connect(self.table_settings_changed)
 
         self.group_box_table_settings.setLayout(wl_layouts.Wl_Layout())
-        self.group_box_table_settings.layout().addWidget(self.checkbox_show_pct, 0, 0)
+        self.group_box_table_settings.layout().addWidget(self.checkbox_show_pct_data, 0, 0)
 
         # Figure Settings
         self.group_box_fig_settings = QGroupBox(self.tr('Figure Settings'), self)
@@ -235,7 +235,7 @@ class Wrapper_Dependency_Parser(wl_layouts.Wl_Wrapper):
             self.main.wl_context_settings_dependency_parser.load_settings(defaults = True)
 
         # Table Settings
-        self.checkbox_show_pct.setChecked(settings['table_settings']['show_pct'])
+        self.checkbox_show_pct_data.setChecked(settings['table_settings']['show_pct_data'])
 
         # Figure Settings
         self.checkbox_show_pos_tags.setChecked(settings['fig_settings']['show_pos_tags'])
@@ -284,7 +284,7 @@ class Wrapper_Dependency_Parser(wl_layouts.Wl_Wrapper):
     def table_settings_changed(self):
         settings = self.main.settings_custom['dependency_parser']['table_settings']
 
-        settings['show_pct'] = self.checkbox_show_pct.isChecked()
+        settings['show_pct_data'] = self.checkbox_show_pct_data.isChecked()
 
     def fig_settings_changed(self):
         settings = self.main.settings_custom['dependency_parser']['fig_settings']
@@ -396,7 +396,7 @@ class Wl_Table_Dependency_Parser(wl_tables.Wl_Table_Data_Search):
 
                 self.enable_updates()
 
-                self.toggle_pct()
+                self.toggle_pct_data()
             except Exception:
                 err_msg = traceback.format_exc()
             finally: